JWTキー生成コマンド追加
