+++ plugins/baser-core/src/BaserCorePlugin.php
@@ -17,6 +17,7 @@ use Authentication\AuthenticationServiceProviderInterface;
 use Authentication\Authenticator\SessionAuthenticator;
 use Authentication\Middleware\AuthenticationMiddleware;
 use BaserCore\Command\ComposerCommand;
+use BaserCore\Command\CreateJwtCommand;
 use BaserCore\Command\CreateReleaseCommand;
 use BaserCore\Command\SetupInstallCommand;
 use BaserCore\Command\SetupTestCommand;
@@ -670,6 +671,7 @@ class BaserCorePlugin extends BcPlugin implements AuthenticationServiceProviderI
         $commands->add('update', UpdateCommand::class);
         $commands->add('create release', CreateReleaseCommand::class);
         $commands->add('setup install', SetupInstallCommand::class);
+        $commands->add('create jwt', CreateJwtCommand::class);
         return $commands;
     }
