+++ plugins/bc-installer/src/Service/InstallationsService.php
@@ -507,27 +507,6 @@ class InstallationsService implements InstallationsServiceInterface
     }
 
     /**
-     * JWTキーを作成する
-     *
-     * @return bool
-     * @noTodo
-     * @checked
-     * @unitTest
-     */
-    public function createJwt()
-    {
-        $command = "openssl genrsa -out " . CONFIG . "jwt.key 1024";
-        exec($command, $out, $code);
-        if($code === 0) {
-            $command = "openssl rsa -in " . CONFIG . "jwt.key -outform PEM -pubout -out " . CONFIG . "jwt.pem";
-            exec($command, $out, $code);
-            return ($code === 0);
-        } else {
-            return false;
-        }
-    }
-
-    /**
      * エディタテンプレート用のアイコン画像をデプロイ
      *
      * @return boolean
